- Expanded single "GlobalEnums" to multiple, more descriptive Enums. - Moved "extract_key_value_cli" from utilities to validator.py (only there it was used). - Moved "add_dataset_source" from core to utilities now that the circular import has been fixed.
